fix seeder ranks

+++ BackEnd/database/migrations/2024_11_30_214348_create_ranks_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('ranks', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->float('total_order_amount');
+            $table->integer('total_order_amount');
             $table->integer('percent_discount');
             $table->timestamps();
         });
